requirements.txt
flask==3.0.3
torch==2.4.1
torchvision==0.19.1
timm==1.0.9
pillow==10.4.0
gunicorn==22.0.0
